% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{calculate_SPEI_MECC}
\alias{calculate_SPEI_MECC}
\title{Function to calculate SPEI}
\usage{
calculate_SPEI_MECC(Date, Precipitation, ETP, scale = 3)
}
\arguments{
\item{Date}{series of dates (daily) to be of the time series.}

\item{Precipitation}{series of daily precipitation (mm)}

\item{ETP}{series of minimum daily potential evapotranspiration (mm), preferably obtained with `calculate_ETP_daily_MECC()`}

\item{scale}{integer, the number of months for accumulation computation}
}
\description{
Function to calculate SPEI
}
