% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_copula_multipliers}
\alias{get_copula_multipliers}
\title{function to calculate the copula's Lagrange multipliers}
\usage{
get_copula_multipliers(data)
}
\arguments{
\item{data}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals. Preferably the output of `get_marginal_distribution()`}
}
\value{
a numeric vector with names and six elements, the Lagrangian multipliers *Lambda_0*,
*Lambda_1*, and *Lambda_2*, *Gamma_1*, *Gamma_2*, and *Lambda_3*
}
\description{
function to calculate the copula's Lagrange multipliers
}
