% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{plot_tr_distribution_comparison}
\alias{plot_tr_distribution_comparison}
\title{function to plot the  changes in the distribution of return period period (Tr) for the domain of events in the reference period.}
\usage{
plot_tr_distribution_comparison(
  tr_dist_comparison,
  data_reference,
  data_study,
  na.rm = F
)
}
\arguments{
\item{tr_dist_comparison}{The output of function `get_distribution_comparison()`. Contains
reference event feature values and the two distributions of TR.}

\item{data_reference}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals for the REFERENCE period of interest. Preferably
the output of `get_marginal_distribution()`.}

\item{data_study}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals for the STUDY period of interest. Preferably
the output of `get_marginal_distribution()`.}

\item{na.rm}{logical. Default is True. NA values are filled with `tidyr::fill(direction = 'down')`}
}
\description{
function to plot the  changes in the distribution of return period period (Tr) for the domain of events in the reference period.
}
